CRUD Producto

+++ src/main/java/TPI/TPI/dao/api/ProductoDaoAPI.java
@@ -4,5 +4,5 @@ import TPI.TPI.Entity.Productos;
 import org.apache.commons.math3.stat.descriptive.summary.Product;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProductoDaoAPI extends CrudRepository<Productos, Long> {
+public interface ProductoDaoAPI extends CrudRepository<Productos, Integer> {
 }
