+++ src/main/java/TPI/TPI/service/api/ProductoServiceAPI.java
@@ -3,5 +3,5 @@ package TPI.TPI.service.api;
 import TPI.TPI.Commons.GenericServiceAPI;
 import TPI.TPI.Entity.Productos;
 
-public interface ProductoServiceAPI extends GenericServiceAPI<Productos, Long> {
+public interface ProductoServiceAPI extends GenericServiceAPI<Productos, Integer> {
 }
